Practice Full Stack dev /requirements.txt
asgiref==3.6.0
Django==3.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
